requirements.txt
# ===========================================
# DNA Methylation Analysis Pipeline
# Python Dependencies
# ===========================================
# Python 3.8+ required

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration management
PyYAML>=5.4.0

# Optional: Progress bars for large datasets
# tqdm>=4.62.0

# Optional: Jupyter notebook support
# jupyter>=1.0.0
# jupyterlab>=3.0.0

# ===========================================
# R Dependencies (install separately)
# ===========================================
# Install via BiocManager in R:
#   DMRcate
#   minfi
#   limma
#   GenomicRanges
#   IlluminaHumanMethylationEPICanno.ilm10b5.hg38
